+++ components/browser/storage-sync/src/main/java/mozilla/components/browser/storage/sync/Connection.kt
@@ -8,11 +8,6 @@ import androidx.annotation.GuardedBy
 import mozilla.appservices.places.PlacesApi
 import mozilla.appservices.places.PlacesReaderConnection
 import mozilla.appservices.places.PlacesWriterConnection
-import mozilla.appservices.sync15.FailureName
-import mozilla.appservices.sync15.FailureReason
-import mozilla.appservices.sync15.SyncTelemetryPing
-import mozilla.components.browser.storage.sync.GleanMetrics.BookmarksSync
-import mozilla.components.browser.storage.sync.GleanMetrics.HistorySync
 import mozilla.components.browser.storage.sync.GleanMetrics.Pings
 import mozilla.components.concept.sync.SyncAuthInfo
 import java.io.Closeable
@@ -39,112 +34,8 @@ internal interface Connection : Closeable {
     fun syncHistory(syncInfo: SyncAuthInfo)
     fun syncBookmarks(syncInfo: SyncAuthInfo)
 
-    // These are implemented as default methods on `Connection` instead of
-    // `RustPlacesConnection` to make testing easier.
-    @Suppress("ComplexMethod", "NestedBlockDepth")
-    fun assembleHistoryPing(ping: SyncTelemetryPing) {
-        ping.syncs.forEach eachSync@{ sync ->
-            sync.failureReason?.let {
-                recordHistoryFailureReason(it)
-                sendHistoryPing()
-                return@eachSync
-            }
-            sync.engines.forEach eachEngine@{ engine ->
-                if (engine.name != "history") {
-                    return@eachEngine
-                }
-                HistorySync.apply {
-                    val base = BaseGleanSyncPing.fromEngineInfo(ping.uid, engine)
-                    uid.set(base.uid)
-                    startedAt.set(base.startedAt)
-                    finishedAt.set(base.finishedAt)
-                    if (base.applied > 0) {
-                        // Since all Sync ping counters have `lifetime: ping`, and
-                        // we send the ping immediately after, we don't need to
-                        // reset the counters before calling `add`.
-                        incoming["applied"].add(base.applied)
-                    }
-                    if (base.failedToApply > 0) {
-                        incoming["failed_to_apply"].add(base.failedToApply)
-                    }
-                    if (base.reconciled > 0) {
-                        incoming["reconciled"].add(base.reconciled)
-                    }
-                    if (base.uploaded > 0) {
-                        outgoing["uploaded"].add(base.uploaded)
-                    }
-                    if (base.failedToUpload > 0) {
-                        outgoing["failed_to_upload"].add(base.failedToUpload)
-                    }
-                    if (base.outgoingBatches > 0) {
-                        outgoingBatches.add(base.outgoingBatches)
-                    }
-                    base.failureReason?.let {
-                        recordHistoryFailureReason(it)
-                    }
-                }
-                sendHistoryPing()
-            }
-        }
-    }
-
     @Suppress("EmptyFunctionBlock")
     fun sendHistoryPing() {}
-
-    // This function is almost identical to `recordHistoryPing`, with additional
-    // reporting for validation problems. Unfortunately, since the
-    // `BookmarksSync` and `HistorySync` metrics are two separate objects, we
-    // can't factor this out into a generic function.
-    @Suppress("ComplexMethod", "NestedBlockDepth")
-    fun assembleBookmarksPing(ping: SyncTelemetryPing) {
-        ping.syncs.forEach eachSync@{ sync ->
-            sync.failureReason?.let {
-                // If the entire sync fails, don't try to unpack the ping; just
-                // report the error and bail.
-                recordBookmarksFailureReason(it)
-                sendBookmarksPing()
-                return@eachSync
-            }
-            sync.engines.forEach eachEngine@{ engine ->
-                if (engine.name != "bookmarks") {
-                    return@eachEngine
-                }
-                BookmarksSync.apply {
-                    val base = BaseGleanSyncPing.fromEngineInfo(ping.uid, engine)
-                    uid.set(base.uid)
-                    startedAt.set(base.startedAt)
-                    finishedAt.set(base.finishedAt)
-                    if (base.applied > 0) {
-                        incoming["applied"].add(base.applied)
-                    }
-                    if (base.failedToApply > 0) {
-                        incoming["failed_to_apply"].add(base.failedToApply)
-                    }
-                    if (base.reconciled > 0) {
-                        incoming["reconciled"].add(base.reconciled)
-                    }
-                    if (base.uploaded > 0) {
-                        outgoing["uploaded"].add(base.uploaded)
-                    }
-                    if (base.failedToUpload > 0) {
-                        outgoing["failed_to_upload"].add(base.failedToUpload)
-                    }
-                    if (base.outgoingBatches > 0) {
-                        outgoingBatches.add(base.outgoingBatches)
-                    }
-                    base.failureReason?.let {
-                        recordBookmarksFailureReason(it)
-                    }
-                    engine.validation?.let {
-                        it.problems.forEach {
-                            remoteTreeProblems[it.name].add(it.count)
-                        }
-                    }
-                }
-                sendBookmarksPing()
-            }
-        }
-    }
 
     @Suppress("EmptyFunctionBlock")
     fun sendBookmarksPing() {}
@@ -186,7 +77,9 @@ internal object RustPlacesConnection : Connection {
     override fun syncHistory(syncInfo: SyncAuthInfo) {
         check(api != null) { "must call init first" }
         val ping = api!!.syncHistory(syncInfo.into())
-        assembleHistoryPing(ping)
+        SyncTelemetry.processHistoryPing(ping) {
+            Pings.historySync.send()
+        }
     }
 
     override fun sendHistoryPing() {
@@ -196,7 +89,9 @@ internal object RustPlacesConnection : Connection {
     override fun syncBookmarks(syncInfo: SyncAuthInfo) {
         check(api != null) { "must call init first" }
         val ping = api!!.syncBookmarks(syncInfo.into())
-        assembleBookmarksPing(ping)
+        SyncTelemetry.processBookmarksPing(ping) {
+            Pings.bookmarksSync.send()
+        }
     }
 
     override fun sendBookmarksPing() {
@@ -208,26 +103,4 @@ internal object RustPlacesConnection : Connection {
         api!!.close()
         api = null
     }
-}
-
-private fun recordHistoryFailureReason(reason: FailureReason) {
-    val metric = when (reason.name) {
-        FailureName.Other, FailureName.Unknown -> HistorySync.failureReason["other"]
-        FailureName.Unexpected, FailureName.Http -> HistorySync.failureReason["unexpected"]
-        FailureName.Auth -> HistorySync.failureReason["auth"]
-        FailureName.Shutdown -> return
-    }
-    val message = reason.message ?: "Unexpected error: ${reason.code}"
-    metric.set(message.take(MAX_FAILURE_REASON_LENGTH))
-}
-
-private fun recordBookmarksFailureReason(reason: FailureReason) {
-    val metric = when (reason.name) {
-        FailureName.Other, FailureName.Unknown -> BookmarksSync.failureReason["other"]
-        FailureName.Unexpected, FailureName.Http -> BookmarksSync.failureReason["unexpected"]
-        FailureName.Auth -> BookmarksSync.failureReason["auth"]
-        FailureName.Shutdown -> return
-    }
-    val message = reason.message ?: "Unexpected error: ${reason.code}"
-    metric.set(message.take(MAX_FAILURE_REASON_LENGTH))
 }
