+++ components/browser/storage-sync/src/test/java/mozilla/components/browser/storage/sync/PlacesHistoryStorageTest.kt
@@ -10,39 +10,23 @@ import mozilla.appservices.places.InternalPanic
 import mozilla.appservices.places.PlacesException
 import mozilla.appservices.places.PlacesReaderConnection
 import mozilla.appservices.places.PlacesWriterConnection
-import mozilla.appservices.sync15.EngineInfo
-import mozilla.appservices.sync15.FailureName
-import mozilla.appservices.sync15.FailureReason
-import mozilla.appservices.sync15.IncomingInfo
-import mozilla.appservices.sync15.OutgoingInfo
-import mozilla.appservices.sync15.SyncInfo
-import mozilla.appservices.sync15.SyncTelemetryPing
-import mozilla.components.browser.storage.sync.GleanMetrics.HistorySync
-import mozilla.components.browser.storage.sync.GleanMetrics.Pings
 import mozilla.components.concept.storage.PageObservation
 import mozilla.components.concept.storage.VisitType
 import mozilla.components.concept.sync.SyncAuthInfo
 import mozilla.components.concept.sync.SyncStatus
-import mozilla.components.service.glean.testing.GleanTestRule
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
 import org.junit.After
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Assert.fail
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import java.util.Date
 
 @RunWith(AndroidJUnit4::class)
 class PlacesHistoryStorageTest {
-    @get:Rule
-    val gleanRule = GleanTestRule(testContext)
-
     private lateinit var history: PlacesHistoryStorage
 
     @Before
@@ -593,387 +577,6 @@ class PlacesHistoryStorageTest {
 
         val error = result as SyncStatus.Error
         assertEquals("test error", error.exception.message)
-    }
-
-    @Test
-    fun `sends history telemetry pings on success`() = runBlocking {
-        val now = Date().asSeconds()
-        val conn = object : Connection {
-            var pingCount = 0
-
-            override fun reader(): PlacesReaderConnection {
-                fail()
-                return mock()
-            }
-
-            override fun writer(): PlacesWriterConnection {
-                fail()
-                return mock()
-            }
-
-            override fun syncHistory(syncInfo: SyncAuthInfo) {
-                assembleHistoryPing(SyncTelemetryPing(
-                    version = 1,
-                    uid = "abc123",
-                    syncs = listOf(
-                        SyncInfo(
-                            at = now,
-                            took = 10000,
-                            engines = listOf(
-                                EngineInfo(
-                                    name = "logins",
-                                    at = now + 5,
-                                    took = 5000,
-                                    incoming = null,
-                                    outgoing = emptyList(),
-                                    failureReason = null,
-                                    validation = null
-                                ),
-                                EngineInfo(
-                                    name = "history",
-                                    at = now,
-                                    took = 5000,
-                                    incoming = IncomingInfo(
-                                        applied = 5,
-                                        failed = 4,
-                                        newFailed = 3,
-                                        reconciled = 2
-                                    ),
-                                    outgoing = listOf(
-                                        OutgoingInfo(
-                                            sent = 10,
-                                            failed = 5
-                                        ),
-                                        OutgoingInfo(
-                                            sent = 4,
-                                            failed = 2
-                                        )
-                                    ),
-                                    failureReason = null,
-                                    validation = null
-                                )
-                            ),
-                            failureReason = null
-                        ),
-                        SyncInfo(
-                            at = now + 10,
-                            took = 5000,
-                            engines = listOf(
-                                EngineInfo(
-                                    name = "history",
-                                    at = now + 10,
-                                    took = 5000,
-                                    incoming = null,
-                                    outgoing = emptyList(),
-                                    failureReason = null,
-                                    validation = null
-                                )
-                            ),
-                            failureReason = null
-                        )
-                    ),
-                    events = emptyList()
-                ))
-            }
-
-            override fun sendHistoryPing() {
-                when (pingCount) {
-                    0 -> {
-                        HistorySync.apply {
-                            assertEquals("abc123", uid.testGetValue())
-                            assertEquals(now, startedAt.testGetValue().asSeconds())
-                            assertEquals(now + 5, finishedAt.testGetValue().asSeconds())
-                            assertEquals(5, incoming["applied"].testGetValue())
-                            assertEquals(7, incoming["failed_to_apply"].testGetValue())
-                            assertEquals(2, incoming["reconciled"].testGetValue())
-                            assertEquals(14, outgoing["uploaded"].testGetValue())
-                            assertEquals(7, outgoing["failed_to_upload"].testGetValue())
-                            assertEquals(2, outgoingBatches.testGetValue())
-                        }
-                    }
-                    1 -> {
-                        HistorySync.apply {
-                            assertEquals("abc123", uid.testGetValue())
-                            assertEquals(now + 10, startedAt.testGetValue().asSeconds())
-                            assertEquals(now + 15, finishedAt.testGetValue().asSeconds())
-                            assertTrue(listOf(
-                                incoming["applied"],
-                                incoming["failed_to_apply"],
-                                incoming["reconciled"],
-                                outgoing["uploaded"],
-                                outgoing["failed_to_upload"],
-                                outgoingBatches
-                            ).none { it.testHasValue() })
-                        }
-                    }
-                    else -> fail()
-                }
-                // We still need to send the ping, so that the counters are
-                // cleared out between calls to `sendHistoryPing`.
-                Pings.historySync.send()
-                pingCount++
-            }
-
-            override fun syncBookmarks(syncInfo: SyncAuthInfo) {
-                fail()
-            }
-
-            override fun sendBookmarksPing() {
-                fail()
-            }
-
-            override fun close() {
-                fail()
-            }
-        }
-        val storage = MockingPlacesHistoryStorage(conn)
-
-        val result = storage.sync(SyncAuthInfo("kid", "token", 123L, "key", "serverUrl"))
-
-        assertTrue(result is SyncStatus.Ok)
-        assertEquals(2, conn.pingCount)
-    }
-
-    @Test
-    fun `sends history telemetry pings on engine failure`() = runBlocking {
-        val now = Date().asSeconds()
-        val conn = object : Connection {
-            var pingCount = 0
-
-            override fun reader(): PlacesReaderConnection {
-                fail()
-                return mock()
-            }
-
-            override fun writer(): PlacesWriterConnection {
-                fail()
-                return mock()
-            }
-
-            override fun syncHistory(syncInfo: SyncAuthInfo) {
-                assembleHistoryPing(SyncTelemetryPing(
-                    version = 1,
-                    uid = "abc123",
-                    syncs = listOf(
-                        SyncInfo(
-                            at = now,
-                            took = 5000,
-                            engines = listOf(
-                                // We should ignore any engines that aren't
-                                // history.
-                                EngineInfo(
-                                    name = "bookmarks",
-                                    at = now + 1,
-                                    took = 1000,
-                                    incoming = null,
-                                    outgoing = emptyList(),
-                                    failureReason = FailureReason(FailureName.Unknown, "Boxes not locked"),
-                                    validation = null
-                                ),
-                                // Multiple history engine syncs per sync isn't
-                                // expected, but it's easier to test the
-                                // different failure types this way, instead of
-                                // creating a top-level `SyncInfo` for each
-                                // one.
-                                EngineInfo(
-                                    name = "history",
-                                    at = now + 2,
-                                    took = 1000,
-                                    incoming = null,
-                                    outgoing = emptyList(),
-                                    failureReason = FailureReason(FailureName.Shutdown),
-                                    validation = null
-                                ),
-                                EngineInfo(
-                                    name = "history",
-                                    at = now + 3,
-                                    took = 1000,
-                                    incoming = null,
-                                    outgoing = emptyList(),
-                                    failureReason = FailureReason(FailureName.Unknown, "Synergies not aligned"),
-                                    validation = null
-                                ),
-                                EngineInfo(
-                                    name = "history",
-                                    at = now + 4,
-                                    took = 1000,
-                                    incoming = null,
-                                    outgoing = emptyList(),
-                                    failureReason = FailureReason(FailureName.Http, code = 418),
-                                    validation = null
-                                )
-                            ),
-                            failureReason = null
-                        ),
-                        // ...But, just in case, we also test multiple top-level
-                        // syncs.
-                        SyncInfo(
-                            at = now + 5,
-                            took = 4000,
-                            engines = listOf(
-                                EngineInfo(
-                                    name = "history",
-                                    at = now + 6,
-                                    took = 1000,
-                                    incoming = null,
-                                    outgoing = emptyList(),
-                                    failureReason = FailureReason(FailureName.Auth, "Splines not reticulated", 999),
-                                    validation = null
-                                ),
-                                EngineInfo(
-                                    name = "history",
-                                    at = now + 7,
-                                    took = 1000,
-                                    incoming = null,
-                                    outgoing = emptyList(),
-                                    failureReason = FailureReason(FailureName.Unexpected, "Kaboom!"),
-                                    validation = null
-                                ),
-                                EngineInfo(
-                                    name = "history",
-                                    at = now + 8,
-                                    took = 1000,
-                                    incoming = null,
-                                    outgoing = emptyList(),
-                                    failureReason = FailureReason(FailureName.Other, "Qualia unsynchronized"), // other
-                                    validation = null
-                                )
-                            ),
-                            failureReason = null
-                        )
-                    ),
-                    events = emptyList()
-                ))
-            }
-
-            override fun sendHistoryPing() {
-                when (pingCount) {
-                    0 -> {
-                        // Shutdown errors shouldn't be reported at all.
-                        assertTrue(listOf(
-                            "other",
-                            "unexpected",
-                            "auth"
-                        ).none { HistorySync.failureReason[it].testHasValue() })
-                    }
-                    1 -> HistorySync.apply {
-                        assertEquals("Synergies not aligned", failureReason["other"].testGetValue())
-                        assertFalse(failureReason["unexpected"].testHasValue())
-                        assertFalse(failureReason["auth"].testHasValue())
-                    }
-                    2 -> HistorySync.apply {
-                        assertEquals("Unexpected error: 418", failureReason["unexpected"].testGetValue())
-                        assertFalse(failureReason["other"].testHasValue())
-                        assertFalse(failureReason["auth"].testHasValue())
-                    }
-                    3 -> HistorySync.apply {
-                        assertEquals("Splines not reticulated", failureReason["auth"].testGetValue())
-                        assertFalse(failureReason["other"].testHasValue())
-                        assertFalse(failureReason["unexpected"].testHasValue())
-                    }
-                    4 -> HistorySync.apply {
-                        assertEquals("Kaboom!", failureReason["unexpected"].testGetValue())
-                        assertFalse(failureReason["other"].testHasValue())
-                        assertFalse(failureReason["auth"].testHasValue())
-                    }
-                    5 -> HistorySync.apply {
-                        assertEquals("Qualia unsynchronized", failureReason["other"].testGetValue())
-                        assertFalse(failureReason["unexpected"].testHasValue())
-                        assertFalse(failureReason["auth"].testHasValue())
-                    }
-                    else -> fail()
-                }
-                // We still need to send the ping, so that the counters are
-                // cleared out between calls to `sendHistoryPing`.
-                Pings.historySync.send()
-                pingCount++
-            }
-
-            override fun syncBookmarks(syncInfo: SyncAuthInfo) {
-                fail()
-            }
-
-            override fun sendBookmarksPing() {
-                fail()
-            }
-
-            override fun close() {
-                fail()
-            }
-        }
-        val storage = MockingPlacesHistoryStorage(conn)
-
-        val result = storage.sync(SyncAuthInfo("kid", "token", 123L, "key", "serverUrl"))
-
-        assertTrue(result is SyncStatus.Ok)
-        assertEquals(6, conn.pingCount)
-    }
-
-    @Test
-    fun `sends history telemetry pings on sync failure`() = runBlocking {
-        val now = Date().asSeconds()
-        val conn = object : Connection {
-            var pingCount = 0
-
-            override fun reader(): PlacesReaderConnection {
-                fail()
-                return mock()
-            }
-
-            override fun writer(): PlacesWriterConnection {
-                fail()
-                return mock()
-            }
-
-            override fun syncHistory(syncInfo: SyncAuthInfo) {
-                assembleHistoryPing(SyncTelemetryPing(
-                    version = 1,
-                    uid = "abc123",
-                    syncs = listOf(
-                        SyncInfo(
-                            at = now,
-                            took = 5000,
-                            engines = emptyList(),
-                            failureReason = FailureReason(FailureName.Unknown, "Synergies not aligned")
-                        )
-                    ),
-                    events = emptyList()
-                ))
-            }
-
-            override fun sendHistoryPing() {
-                when (pingCount) {
-                    0 -> HistorySync.apply {
-                        assertEquals("Synergies not aligned", failureReason["other"].testGetValue())
-                        assertFalse(failureReason["unexpected"].testHasValue())
-                        assertFalse(failureReason["auth"].testHasValue())
-                    }
-                    else -> fail()
-                }
-                // We still need to send the ping, so that the counters are
-                // cleared out between calls to `sendHistoryPing`.
-                Pings.historySync.send()
-                pingCount++
-            }
-
-            override fun syncBookmarks(syncInfo: SyncAuthInfo) {
-                fail()
-            }
-
-            override fun sendBookmarksPing() {
-                fail()
-            }
-
-            override fun close() {
-                fail()
-            }
-        }
-        val storage = MockingPlacesHistoryStorage(conn)
-
-        val result = storage.sync(SyncAuthInfo("kid", "token", 123L, "key", "serverUrl"))
-
-        assertTrue(result is SyncStatus.Ok)
-        assertEquals(1, conn.pingCount)
     }
 
     @Test(expected = InternalPanic::class)
